Seed refactor

+++ server/database/models/TransferChore.js
@@ -3,7 +3,7 @@ const db = require('../db');
 
 const TransferChore = db.define('transferChore', {
   price: {
-    type: Sequelize.DECIMAL(10, 8),
+    type: Sequelize.INTEGER,
     allowNull: false,
   },
   status: {
